Extract composable of semantic colors so that we can have a preview in Android Studio and split into 2 columns
